// src/app/maps/interfaces/places.interface.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
  type:        string;
  query:       number[];
  features:    Feature[];
  attribution: string;
}

export interface Feature {
  id:            string;
  type:          string;
  place_type:    string[];
  relevance:     number;
  properties:    Properties;
  text_es:       string;
  language_es:   string;
  place_name_es: string;
  text:          string;
  language:      string;
  place_name:    string;
  center:        number[];
  geometry:      Geometry;
  context:       Context[];
}

export interface Context {
  id:           string;
  mapbox_id:    string;
  text_es:      string;
  text:         string;
  wikidata?:    string;
  language_es?: string;
  language?:    string;
  short_code?:  string;
}

export interface Geometry {
  coordinates: number[];
  type:        string;
}

export interface Properties {
  foursquare: string;
  landmark:   boolean;
  wikidata:   string;
  address:    string;
  category:   string;
  maki:       string;
}